fix: fix js mapping and cleanup
