+++ qtmonaco/qtmonaco/monaco_page.py
@@ -7,4 +7,4 @@ logger = logging.getLogger(__name__)
 
 class MonacoPage(QWebEnginePage):
     def javaScriptConsoleMessage(self, level, message, line, source):
-        logger.debug(f"[JS Console] {level.name} at line {line} in {source}: {message}")
+        print(f"[JS Console] {level.name} at line {line} in {source}: {message}")
